Update contact details listings to include email/phone
