+++ cms/core/wagtail_hooks.py
@@ -1,8 +1,7 @@
 from wagtail import hooks
 from wagtail.snippets.models import register_snippet
-from wagtail.snippets.views.snippets import SnippetViewSet
 
-from cms.core.models import ContactDetails
+from cms.core.viewsets import ContactDetailsViewSet
 
 
 @hooks.register("register_icons")
@@ -19,18 +18,6 @@ def register_icons(icons: list[str]) -> list[str]:
         "identity.svg",
         "news.svg",
     ]
-
-
-class ContactDetailsViewSet(SnippetViewSet):
-    """A snippet viewset for ContactDetails.
-
-    See:
-     - https://docs.wagtail.org/en/stable/topics/snippets/registering.html
-     - https://docs.wagtail.org/en/stable/topics/snippets/customizing.html#icon
-    """
-
-    model = ContactDetails
-    icon = "identity"
 
 
 register_snippet(ContactDetailsViewSet)
